migrated ButtonTheme HOC to hook
