+++ src/components/Button/useButtonThemeAnim.ts
@@ -7,23 +7,23 @@ import type { ButtonTheme } from './ButtonNew';
 export const useButtonThemeAnim = (
   disabled: boolean | undefined,
   isElevatedMode: boolean,
-  theme: ButtonTheme
+  elevation: ButtonTheme['elevation']
 ) => {
   // animation should come from ButtonTheme
-  const { animation } = useInternalTheme();
-  const { current: elevation } = useRef<Animated.Value>(
-    new Animated.Value(disabled ? 0 : theme.elevation.initial)
+  const { animation } = useInternalTheme(undefined);
+  const { current: elevationRef } = useRef<Animated.Value>(
+    new Animated.Value(disabled ? 0 : elevation.initial)
   );
 
   useEffect(() => {
-    elevation.setValue(disabled ? 0 : theme.elevation.initial);
-  }, [disabled, elevation, theme]);
+    elevationRef.setValue(disabled ? 0 : elevation.initial);
+  }, [disabled, elevationRef, elevation]);
 
   const animatePressIn = () => {
     if (isElevatedMode) {
       const { scale } = animation;
-      Animated.timing(elevation, {
-        toValue: theme.elevation.active,
+      Animated.timing(elevationRef, {
+        toValue: elevation.active,
         duration: 200 * scale,
         useNativeDriver: true,
       }).start();
@@ -33,8 +33,8 @@ export const useButtonThemeAnim = (
   const animatePressOut = () => {
     if (isElevatedMode) {
       const { scale } = animation;
-      Animated.timing(elevation, {
-        toValue: theme.elevation.initial,
+      Animated.timing(elevationRef, {
+        toValue: elevation.initial,
         duration: 150 * scale,
         useNativeDriver: true,
       }).start();
@@ -42,7 +42,7 @@ export const useButtonThemeAnim = (
   };
 
   return {
-    elevationAnim: elevation,
+    elevationAnim: elevationRef,
     animatePressIn,
     animatePressOut,
   };
